feat: now libertybans should support alt registration per server. (#7)
* feat: now libertybans should support alt registration per server.

This means that now you can choose whether all servers will register alts or only the selected ones. For proxies, you need to provide the names of the servers that you have in your configs, and for backends just enable or disable a setting in config.

* feat: now libertybans should support alt registration per server.

This means that now you can choose whether all servers will register alts or only the selected ones. For proxies, you need to provide the names of the servers that you have in your configs, and for backends just enable or disable a setting in config.

* refactor

* feat: should be working now [not tested yet]

* fix: moved check outside of db query.

* feat: added more precise explanation

* Accomodate unrecorded IP address in applicability calculation

Plus a couple tweaks to code style

* refactoring

* refactoring

* Add sonatype as repository

* fix: extra ")" in ConfigKey hould-register-on-connection

---------

Co-authored-by: Snake <lluisponsvidal@gmail.com>
Co-authored-by: A248 <theanandbeh@gmail.com>